+++ front/src/components/table/useFileDataGrid.jsx
@@ -1,12 +1,39 @@
-import {useState} from 'react'
-import {GridActionsCellItem} from '@mui/x-data-grid'
-import EditIcon from '@mui/icons-material/Edit'
 import DeleteIcon from '@mui/icons-material/Delete'
+import EditIcon from '@mui/icons-material/Edit'
 import Chip from '@mui/material/Chip'
+import {GridActionsCellItem} from '@mui/x-data-grid'
+import {useQuery} from '@tanstack/react-query'
+import {useState} from 'react'
+import {fetchGetFiles} from '../../utils/fetch'
 
 function useFileDataGrid() {
     const [rowData, setRowData] = useState({})
     const [open, setOpen] = useState(false)
+
+    const token = localStorage.getItem('BBStorage_token')
+
+    const testGetFiles = useQuery({
+        queryKey: ['GetFiles'],
+        queryFn: fetchGetFiles(token),
+        enabled: !!token,
+    })
+
+    // const {mutate} = useMutation({
+    //     mutationKey: ['updateTitle'],
+    //     mutationFn: (variables) => fetchUpdateFile(variables),
+    //     onSuccess: () => {
+    //         // TODO : Alert "Document mis à jour"
+    //     },
+    //     onError: (error) => {
+    //         // TODO : Alert "Erreur lors de la mise à jour"
+    //     },
+    // })
+
+    // TODO : Mettre en place composant pour récupérer les informations de stockage
+    // const {total_weight_files, total_storage_capacity} = data
+    // const storagePercentage =
+    //     (total_weight_files * 100) / total_storage_capacity
+    const storagePercentage = 50
 
     function handleDeleteClick(id) {
         //TODO : Mettre une modal de confirmation
@@ -97,7 +124,7 @@ function useFileDataGrid() {
         },
     ]
 
-    return {open, setOpen, columns, rowData}
+    return {open, setOpen, columns, rowData, storagePercentage, testGetFiles}
 }
 
 export default useFileDataGrid
